Decoder functions merged into single file
