+++ neutral_example_implentation/device_function_simulator.c
@@ -1,5 +1,8 @@
 #include "device_function_simulator.h"
-#include "getch.h"
+
+#include <termios.h>
+#include <unistd.h>
+#include <fcntl.h>
 
 // Device parameters (to be stored in Flash/EEPROM)
 uint16_t TokenCount = 1;
@@ -7,13 +10,26 @@ bool PAYGEnabled = true;
 uint32_t ActiveUntil = 0;
 uint32_t TokenEntryLockedUntil = 0;
 // WARNING: THIS SECRET KEY AND STARTING CODE IS ONLY HERE AS AN EXAMPLE AND SHOULD NEVER BE USED IN PRODUCTION
-// uint32_t StartingCode = 123456789;
+//uint32_t StartingCode = 123456789;
 //char SECRET_KEY[16] = {0xa2, 0x9a, 0xb8, 0x2e, 0xdc, 0x5f, 0xbb, 0xc4, 0x1e, 0xc9, 0x53, 0xf, 0x6d, 0xac, 0x86, 0xb1};
 char SECRET_KEY[16] = {...};
 
 
 // This is just for displaying the activation status in the simulater WaitForTokenEntry function
 extern bool IsActive(void);
+
+int getch() {
+    struct termios oldt,
+    newt;
+    int            ch;
+    tcgetattr( STDIN_FILENO, &oldt );
+    newt = oldt;
+    newt.c_lflag &= ~( ICANON | ECHO );
+    tcsetattr( STDIN_FILENO, TCSANOW, &newt );
+    ch = getchar();
+    tcsetattr( STDIN_FILENO, TCSANOW, &oldt );
+    return ch;
+}
 
 void BlinkRedLED(int NumberOfBlinks) {
     printf("\nRed LED Blinked %d times!\n", NumberOfBlinks);
@@ -29,11 +45,11 @@ int GetKeyPressed() {
         case '*':
             return STAR_KEY;
             break;
-            
+
         case '#':
             return HASH_KEY;
             break;
-            
+
         default:
             return (int) (this_char - '0'); // this_char is now an int
             break;
