+++ neutral_example_implentation/device_function_simulator.h
