--- neutral_example_implentation/device_simulator/getch.c
@@ -1,19 +1,0 @@
-#include "getch.h"
-
-#include <termios.h>
-#include <unistd.h>
-#include <fcntl.h>
-
-int getch() {
-    struct termios oldt,
-    newt;
-    int            ch;
-    tcgetattr( STDIN_FILENO, &oldt );
-    newt = oldt;
-    newt.c_lflag &= ~( ICANON | ECHO );
-    tcsetattr( STDIN_FILENO, TCSANOW, &newt );
-    ch = getchar();
-    tcsetattr( STDIN_FILENO, TCSANOW, &oldt );
-    return ch;
-}
-
