--- neutral_example_implentation/device_simulator/getch.h
@@ -1,8 +1,0 @@
-#ifndef getch_h
-#define getch_h
-
-#include <stdio.h>
-
-int getch(void);
-
-#endif
