--- neutral_example_implentation/opaygo_decoder/opaygo_value_utils.h
@@ -1,11 +1,0 @@
-#ifndef opaygo_value_utils_h
-#define opaygo_value_utils_h
-
-#include <stdio.h>
-#include <stdint.h>
-
-int DecodeBase(int StartingCodeBase, int TokenBase);
-int GetTokenBase(uint32_t Token);
-uint32_t PutBaseInToken(uint32_t Token, int TokenBase);
-
-#endif
