--- neutral_example_implentation/opaygo_decoder/restricted_digit_set_mode.h
@@ -1,10 +1,0 @@
-#ifndef restricted_digit_set_mode_h
-#define restricted_digit_set_mode_h
-
-#include <stdio.h>
-#include <stdbool.h>
-#include <stdint.h>
-
-uint32_t ConvertFromFourDigitToken(uint64_t FourDigitToken);
-
-#endif
